requirements.txt
editdistance==0.5.3
elasticsearch>=6.0.0,<7.0.0
tqdm
stanfordnlp
nltk
ujson
gdown
glob2
Jinja2
joblib
jsonlines
numpy
pathlib2
prettytable
regex
scipy
scikit-learn
spacy==2.2.3
tagme==0.1.3
unicodecsv
urllib3
tensorflow-gpu==1.14.0
